+++ test/assets/for_loop.lua
@@ -1,0 +1,10 @@
+-- Count up
+local count = 5
+for i=0, count do
+	print('loop='..i)
+end
+
+-- Count by 2s
+for j=2, count*2, 2 do
+	print('jump='..j)
+end