+++ test/assets/hello_world.lua
@@ -1,0 +1,9 @@
+-- This is why we're here.
+local str = 'Hello, world'
+print(str)
+
+-- Testing variadic args in print.
+print(1, 2, 3, nil, true)
+
+-- Empty variadic args resolves to nil.
+print()