+++ test/lualib_test.dart
@@ -1,1 +1,3 @@
+local str = 'Hello, world!'
+print(str)